+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- pot::get_pot_conf_list(): ignore alias network type, not useful for the analysis
 
 ## [0.4.3] 2020-03-08
 ### Fixed